ocr/requirements.txt
pdf2image==1.16.0
Pillow==9.0.0
pytesseract==0.3.8
python-magic==0.4.24
